+++ pkg/analyzer/test/src/diagnostics/augmentation_without_declaration_test.dart
@@ -65,6 +65,24 @@ augment class A {
     ]);
   }
 
+  test_class_field_static() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+class A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment class A {
+  augment static int foo = 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
   test_class_getter() async {
     newFile('$testPackageLibPath/a.dart', r'''
 import augment 'test.dart';
@@ -83,6 +101,24 @@ augment class A {
     ]);
   }
 
+  test_class_getter_static() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+class A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment class A {
+  augment static int get foo => 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
   test_class_method() async {
     newFile('$testPackageLibPath/a.dart', r'''
 import augment 'test.dart';
@@ -95,6 +131,24 @@ augment library 'a.dart';
 
 augment class A {
   augment void foo() {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_class_method_static() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+class A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment class A {
+  augment static void foo() {}
 }
 ''', [
       error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
@@ -137,6 +191,282 @@ augment class A {
     ]);
   }
 
+  test_class_setter_static() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+class A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment class A {
+  augment static set foo(int _) {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_enum() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment enum A {}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 27, 7),
+    ]);
+  }
+
+  test_enum_constructor() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+enum A {v}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment enum A {;
+  augment const A.named();
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_enum_field() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+enum A {v}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment enum A {;
+  augment final int foo = 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_enum_getter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+enum A {v}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment enum A {;
+  augment int get foo => 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_enum_method() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+enum A {v}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment enum A {;
+  augment void foo() {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_enum_setter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+enum A {v}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment enum A {;
+  augment set foo(int _) {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_extension() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension A {}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 27, 7),
+    ]);
+  }
+
+  test_extension_getter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension A on int {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension A {
+  augment int get foo => 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 51, 7),
+    ]);
+  }
+
+  test_extension_method() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension A on int {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension A {
+  augment void foo() {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 51, 7),
+    ]);
+  }
+
+  test_extension_setter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension A on int {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension A {
+  augment set foo(int _) {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 51, 7),
+    ]);
+  }
+
+  test_extensionType() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension type A(int it) {}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 27, 7),
+    ]);
+  }
+
+  test_extensionType_constructor() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension type A(int it) {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension type A(int it) {
+  augment A.named() : this(0);
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 64, 7),
+    ]);
+  }
+
+  test_extensionType_getter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension type A(int it) {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension type A(int it) {
+  augment int get foo => 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 64, 7),
+    ]);
+  }
+
+  test_extensionType_method() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension type A(int it) {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension type A(int it) {
+  augment void foo() {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 64, 7),
+    ]);
+  }
+
+  test_extensionType_setter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+extension type A(int it) {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment extension type A(int it) {
+  augment set foo(int _) {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 64, 7),
+    ]);
+  }
+
   test_mixin() async {
     newFile('$testPackageLibPath/a.dart', r'''
 import augment 'test.dart';
@@ -148,6 +478,96 @@ augment library 'a.dart';
 augment mixin A {}
 ''', [
       error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 27, 7),
+    ]);
+  }
+
+  test_mixin_field() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+mixin A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment mixin A {
+  augment int foo = 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_mixin_getter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+mixin A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment mixin A {
+  augment int get foo => 0;
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_mixin_method() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+mixin A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment mixin A {
+  augment void foo() {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
+    ]);
+  }
+
+  test_mixin_method_valid() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+mixin A {
+  void foo() {}
+}
+''');
+
+    await assertNoErrorsInCode(r'''
+augment library 'a.dart';
+
+augment mixin A {
+  augment void foo() {}
+}
+''');
+  }
+
+  test_mixin_setter() async {
+    newFile('$testPackageLibPath/a.dart', r'''
+import augment 'test.dart';
+
+mixin A {}
+''');
+
+    await assertErrorsInCode(r'''
+augment library 'a.dart';
+
+augment mixin A {
+  augment set foo(int _) {}
+}
+''', [
+      error(CompileTimeErrorCode.AUGMENTATION_WITHOUT_DECLARATION, 47, 7),
     ]);
   }