+++ pkg/meta/CHANGELOG.md
@@ -31,6 +31,8 @@
   ```
 
   See <https://github.com/dart-lang/sdk/issues/52965> for more information.
+- Introduce `TargetKind.optionalParameter`, to indicate that an annotation is
+  valid on any optional parameter declaration.
 
 ## 1.14.0